+++ app/Http/Controllers/AdminController/AdminAbsenController.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
 
+namespace App\Http\Controllers\AdminController;
+
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Absen;
 use App\Models\jurusan;
