+++ app/Http/Controllers/AdminController/AdminLaporanController.php
@@ -1,0 +1,97 @@
+<?php
+
+
+namespace App\Http\Controllers\AdminController;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+use App\Models\laporan;
+use Illuminate\Support\Facades\Session;
+
+class AdminLaporanController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        $data = laporan::orderBy('created_at','desc')->get();  
+        return view('admin.laporan_admin', compact('data'));
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        $laporan=laporan::find($id);
+        return view('admin.show_laporan_admin', compact('laporan'));
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($id)
+    {
+        //
+    }
+
+    public function hapus()
+    {
+        laporan::truncate();
+        Session::flash('hapus', "Berhasil Menghapus !!");
+        return redirect('/laporan-admin');
+    }
+}
