+++ app/Http/Controllers/AdminController/DashboardController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\AdminController;
 
+use App\Http\Controllers\Controller;
 use App\Models\Absen;
 use App\Models\jurusan;
 use Illuminate\Http\Request;
