+++ app/Http/Controllers/SiswaController/AbsenController.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
 
+namespace App\Http\Controllers\SiswaController;
+
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Absen;
 use App\Models\dokumentasi;
