+++ routes/web.php
@@ -3,7 +3,7 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\LoginController;
-use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\AdminController\DashboardController;
 use App\Http\Controllers\DashboardSiswaController;
 use App\Http\Controllers\RegisterController;
 use App\Http\Controllers\SiswaController;
